requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
face-recognition==1.3.0
pillow==10.1.0
deepface==0.0.79
tf-keras==2.15.0
tensorflow==2.15.0
sounddevice==0.4.6
scipy==1.11.4
pyttsx3==2.90
SpeechRecognition==3.10.0
openai==1.3.0
ecdsa==0.18.0
